Support for equal font even when size shrinks
This huge commit addresses several issues

1) Off/On selection fonts are used as expected
2) If the text in some label is too long, the control will adjust font
size AND (!!!) will ensure that other labels are also using the same
font size. I think it is great, because you can really specify desired
font size, which will be used on bigger devices (iPad, iPhone +) and
the control will lower it gracefully for smaller devices
3) The feature above can be turned off if really needed
4) Control is now based on auto layout. This meant that dividerWidth
(and color) had to be moved to the segment appearance to ensure that
they are set beforehand.

Things not yet implemented:

1) Removing segment doesn't work. Will fix it by reassigning
constraints as needed.
2) Changing appearance on the go will not update constraints. I will
get down to this a bit later (although if someone needs it please
submit PR or raise an issue)